Consolidate Pi tool mapping tests
- Preserve legacy tool inputs while validating authoritative file
  details
- Cover successful create and rename results without failed entries
